feat: prefetch test
