+++ src/components/viewAll/PostList.tsx
@@ -66,7 +66,7 @@ function PostList({ queryKey, queryFn, sortBy }: PostListProps) {
     queryFn,
     initialPageParam: undefined as undefined | QueryDocumentSnapshot<DocumentData, DocumentData>,
     getNextPageParam: (lastPage) => {
-      if (lastPage.length === 0 || lastPage.length < 4) {
+      if (lastPage.length === 0) {
         return undefined;
       }
       return lastPage[lastPage.length - 1];
@@ -281,26 +281,20 @@ function PostList({ queryKey, queryFn, sortBy }: PostListProps) {
         )}
       </St.ContentsWrapper>
       <St.MoreContentWrapper>
-        {isFetchingNextPage ? (
+        {/* {isFetchingNextPage ? (
           <Loader />
         ) : hasNextPage ? (
           <button onClick={() => fetchNextPage()}>더 보기</button>
         ) : (
           <p>모든 데이터를 가져왔습니다.</p>
-        )}
-        {/*
-        
-        //전체length === 보여준 length 같을때 ==> 더보기 안보이게 처리
+        )} */}
         {isFetchingNextPage ? (
           <Loader />
-        ) : hasNextPage ? (z
-          <>
-            <button onClick={() => fetchNextPage()}>더 보기</button>
-            <p>{hasNextPage}</p>
-          </>
+        ) : hasNextPage && posts?.length === 2 ? (
+          <button onClick={() => !isFetchingNextPage && fetchNextPage()}>더 보기</button>
         ) : (
           <p>모든 데이터를 가져왔습니다.</p>
-        )} */}
+        )}
       </St.MoreContentWrapper>
     </St.MainSubWrapper>
   );
