+++ src/components/viewAll/style.ts
@@ -4,12 +4,13 @@ import theme from '../../styles/theme';
 
 const ViewAllContainer = styled.div`
   width: 1000px;
+  min-height: 1544px;
 `;
 
 /*유저 등록 데이터 (친환경노하우~습관인증) */
 const MainSubWrapper = styled.div`
   width: 100%;
-  /* margin: auto; */
+  min-height: 485px;
   margin-bottom: 150px; //58에서 변경
 `;
