+++ src/layouts/navbar/NavBar.tsx
@@ -7,11 +7,9 @@ import usePreviousPathname from '../../util/usePreviousPathname';
 import AuthNavBar from './AuthNavBar';
 import St, { LogoContainer } from './style';
 import { useQueryClient } from '@tanstack/react-query';
-import { getAdminPosts } from '../../api/homeApi';
-import { getAdminPostList, getCategoryPosts } from '../../api/pageListApi';
+import { getAdminPostList } from '../../api/pageListApi';
 import { DocumentData, QueryDocumentSnapshot } from 'firebase/firestore';
 import { QUERY_KEYS } from '../../query/keys';
-import { getAllUsers } from '../../api/authApi';
 
 function NavBar() {
   const [isAuthToggleOpen, setIsAuthToggleOpen] = useState(false);
@@ -36,37 +34,37 @@ function NavBar() {
   const queryClient = useQueryClient(); // queryClient 사용
 
   //잘작동 (admin 한정)
-  // const handleHover = async () => {
-  //   queryClient.prefetchInfiniteQuery({
-  //     queryKey: [QUERY_KEYS.ADMIN],
-  //     queryFn: getAdminPostList,
-  //     initialPageParam: undefined as undefined | QueryDocumentSnapshot<DocumentData, DocumentData>,
-  //     staleTime: 60000
-  //   });
-  // };
-
   const handleHover = async () => {
-    const queriesToPrefetch = [
-      { queryKey: QUERY_KEYS.ADMIN, queryFn: getAdminPostList },
-      { queryKey: QUERY_KEYS.KNOWHOW, queryFn: getCategoryPosts('knowHow') },
-      { queryKey: QUERY_KEYS.RECOMMEND, queryFn: getCategoryPosts('recommendation') },
-      { queryKey: QUERY_KEYS.SHARE, queryFn: getCategoryPosts('sharing') },
-      { queryKey: QUERY_KEYS.HABIT, queryFn: getCategoryPosts('habit') },
-      { queryKey: QUERY_KEYS.TOTAL, queryFn: getCategoryPosts('total') },
-      { queryKey: QUERY_KEYS.NOCATEGORY, queryFn: getCategoryPosts('noCategory') }
-
-      // 다른 queryKey와 queryFn을 추가할 수 있습니다.
-    ];
-
-    for (const { queryKey, queryFn } of queriesToPrefetch) {
-      await queryClient.prefetchInfiniteQuery({
-        queryKey: [queryKey],
-        queryFn: queryFn,
-        initialPageParam: undefined as undefined | QueryDocumentSnapshot<DocumentData, DocumentData>,
-        staleTime: 60000
-      });
-    }
+    queryClient.prefetchInfiniteQuery({
+      queryKey: [QUERY_KEYS.ADMIN],
+      queryFn: getAdminPostList,
+      initialPageParam: undefined as undefined | QueryDocumentSnapshot<DocumentData, DocumentData>,
+      staleTime: 60000
+    });
   };
+
+  // const handleHover = async () => {
+  //   const queriesToPrefetch = [
+  //     { queryKey: QUERY_KEYS.ADMIN, queryFn: getAdminPostList },
+  //     { queryKey: QUERY_KEYS.KNOWHOW, queryFn: getCategoryPosts('knowHow') },
+  //     { queryKey: QUERY_KEYS.RECOMMEND, queryFn: getCategoryPosts('recommendation') },
+  //     { queryKey: QUERY_KEYS.SHARE, queryFn: getCategoryPosts('sharing') },
+  //     { queryKey: QUERY_KEYS.HABIT, queryFn: getCategoryPosts('habit') },
+  //     { queryKey: QUERY_KEYS.TOTAL, queryFn: getCategoryPosts('total') },
+  //     { queryKey: QUERY_KEYS.NOCATEGORY, queryFn: getCategoryPosts('noCategory') }
+
+  //     // 다른 queryKey와 queryFn을 추가할 수 있습니다.
+  //   ];
+
+  //   for (const { queryKey, queryFn } of queriesToPrefetch) {
+  //     await queryClient.prefetchInfiniteQuery({
+  //       queryKey: [queryKey],
+  //       queryFn: queryFn,
+  //       initialPageParam: undefined as undefined | QueryDocumentSnapshot<DocumentData, DocumentData>,
+  //       staleTime: 60000
+  //     });
+  //   }
+  // };
 
   // useEffect(() => {
   //   if (prevPathName === '/write') {
